requirements.txt
gTTS
python-docx